// web/src/api/models/collection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ApeRAG API
 * ApeRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CollectionConfig } from './collection-config';

/**
 * Collection is a collection of documents
 * @export
 * @interface Collection
 */
export interface Collection {
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    'type'?: string;
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    'description'?: string;
    /**
     * 
     * @type {CollectionConfig}
     * @memberof Collection
     */
    'config'?: CollectionConfig;
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    'status'?: CollectionStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    'created'?: string;
    /**
     * 
     * @type {string}
     * @memberof Collection
     */
    'updated'?: string;
    /**
     * Whether the collection is published to marketplace
     * @type {boolean}
     * @memberof Collection
     */
    'is_published'?: boolean;
    /**
     * Publication time, null when not published
     * @type {string}
     * @memberof Collection
     */
    'published_at'?: string;
}

export const CollectionStatusEnum = {
    ACTIVE: 'ACTIVE',
    INACTIVE: 'INACTIVE',
    DELETED: 'DELETED'
} as const;

export type CollectionStatusEnum = typeof CollectionStatusEnum[keyof typeof CollectionStatusEnum];